feat: Add get:jsonbytesize script to package.json and create kbsize.ts file
This commit adds a new script "get:jsonbytesize" to the package.json file. The script uses the newly created kbsize.ts file to calculate the size of a given file in kilobytes (KB), megabytes (MB), and gigabytes (GB). The calculated sizes are displayed in a table format using console.table().